Move all callbacks out of DataDocDash.py
