+++ datadoc/frontend/callbacks/Callbacks.py
@@ -2,6 +2,7 @@ import logging
 from typing import Any, Dict, List, Optional, Tuple
 from pydantic import ValidationError
 from datadoc_model import Model
+from dash import Dash, Output, Input, ctx, ALL
 
 import datadoc.state as state
 from datadoc.Enums import SupportedLanguages
@@ -16,6 +17,9 @@ from datadoc.frontend.DisplayDataset import (
 from datadoc.frontend.DisplayVariables import (
     MULTIPLE_LANGUAGE_VARIABLES_METADATA,
     VariableIdentifiers,
+)
+from datadoc.frontend.Builders import (
+    DATASET_METADATA_INPUT,
 )
 
 logger = logging.getLogger(__name__)
@@ -166,3 +170,68 @@ def update_variable_table_language(
         )
     logger.debug(f"Updated variable table language: {language.name}")
     return new_data, False, ""
+
+
+def register_callbacks(app: Dash) -> None:
+    @app.callback(
+        Output("progress-bar", "value"),
+        Output("progress-bar", "label"),
+        Input({"type": DATASET_METADATA_INPUT, "id": ALL}, "value"),
+        Input("variables-table", "data"),
+    )
+    def callback_update_progress(value, data) -> Tuple[int, str]:
+        completion = state.metadata.percent_complete
+        return completion, f"{completion}%"
+
+    @app.callback(
+        Output("success-message", "is_open"),
+        Input("save-button", "n_clicks"),
+        prevent_initial_call=True,
+    )
+    def callback_save_metadata_file(n_clicks):
+        if n_clicks and n_clicks > 0:
+            # Write the final completion percentage to the model
+            state.metadata.meta.percentage_complete = state.metadata.percent_complete
+            state.metadata.write_metadata_document()
+            return True
+        else:
+            return False
+
+    @app.callback(
+        Output(
+            {"type": DATASET_METADATA_INPUT, "id": ALL},
+            "value",
+        ),
+        Input("language-dropdown", "value"),
+    )
+    def callback_change_language(language: str):
+        return update_dataset_metadata_language(SupportedLanguages(language))
+
+    @app.callback(
+        Output("dataset-validation-error", "is_open"),
+        Output("dataset-validation-explanation", "children"),
+        Input({"type": DATASET_METADATA_INPUT, "id": ALL}, "value"),
+        prevent_initial_call=True,
+    )
+    def callback_accept_dataset_metadata_input(value: Any) -> Tuple[bool, str]:
+        # Get the ID of the input that changed. This MUST match the attribute name defined in DataDocDataSet
+        return accept_dataset_metadata_input(
+            ctx.triggered[0]["value"], ctx.triggered_id["id"]
+        )
+
+    @app.callback(
+        Output("variables-table", "data"),
+        Output("variables-validation-error", "is_open"),
+        Output("variables-validation-explanation", "children"),
+        Input("variables-table", "data"),
+        Input("variables-table", "data_previous"),
+        Input("language-dropdown", "value"),
+        prevent_initial_call=True,
+    )
+    def callback_variable_table(
+        data: List[Dict], data_previous: List[Dict], language: str
+    ) -> Tuple[List[Dict], bool, str]:
+        if ctx.triggered_id == "language-dropdown":
+            return update_variable_table_language(data, SupportedLanguages(language))
+        else:
+            return accept_variable_metadata_input(data, data_previous)
